NavBar Beginnings
Add Deal and Fetch Deals are now in the NavBar

Buttons are dynamic, form buttons are always blue, sign in, sign out, form submit.

Other buttons are white with a blue border and gradually change to blue on hover.

Add a Deal, Deal List and Deal Editor are all now in the same area on the page and only show when they are selected. If you select Add a deal and Fetch Deals they are no longer stacked on the page.